[SBI-1105] Add Double Click selection functionality to Search Modal
This change will allow for search modal results lists to have an option to add 'onDoubleClick' functionality to each row. To do this, there needs to be a new rowTemplate defined, adding the property "ng-dblclick=\"grid.appScope.onDblClickRow(row.entity)\""
to the existing rowTemplate.

I added the default rowTemplate as a comment, but it can also be found in the UI-Grid docs here: https://github.com/angular-ui/ui-grid/blob/v3.0.7/src/templates/ui-grid/ui-grid-row.html

The handling of what happens when double clicking results is outside of the directive, and needs to be linked via adding the property: on-dbl-click-row="{FunctionName}" to the search-modal view object.


Former-commit-id: 4aa29d35d744ecd6b759091f6a8b86f41e4e6103