[AFDP-3910] Port 'Download All As Zip' functionality from Paho to Core
Created single directive to handle the downloading as zip. The existing backend code will name the file acm-{id}-ROOT.zip, which can be altered if desired. Paho has an example of changing the default name based on the title of the case-file.

The container information for Task objects isn't stored the same way as the Complaint, Case-file, and Doc Repo objects. There will need to be a new way of getting the root folder id to pass to the compressor API.


Former-commit-id: 77ca9515bc71f8b59fe9cd32b8b9b7067b0d9d2e