AFDP-4015
Unified the configuration of closed state for 'CaseFile' and 'Complaint'.
The configuration of closed state is defined by 'casefile.status.closed'
and `complaint.status.closed' respectively.

This avoids setting properties in 2 different set of files for the event
listeners and entity handlers, which is both error prone and duplication
of the same values in 2 different places.


Former-commit-id: ae0b3570f5fbeec787ea8ba911c0378197f9319f