AFDP-4477
In case `AcmOutlookFolderCreator` cannot be found in the databese, the
following fallback sequence is implemented:

1. try to access folder with the Outlook credentials stored in the
Outlook integration configuration;
2. if the folder can be accessed, add entry to the database, and returned
the instance of `AcmOutlookFolderCreator`
3. if the folder cannot be accessed, an attempt is made to recreate the
folder. In case the folder exists, but was not accessible due to bad
credentials, or other reason, folder creation will fail, and the
administrator should be informed.If it doesn't exist, and credentials are
valid, it will be recreated.


Former-commit-id: f04a39d8466315eb68a58a566ad12d9aaf06d905