AFDP-4290
Refactored the functionality for downloading attachemnts on the front end,
the functionality is factored out from the controller.

Also fixed the problem with recreation of Outlook folders, the folder ID
was not stored in the container object, due to the fact that the update
was executed out of transaction, which is changed with this commit.


Former-commit-id: 9fa7139c672c54f1ae13efab151d48f52c09515b