AFDP-4813
Calendar integration configuration is kept in memory now, and loaded only
once on startup, and when updated it is written on the filesystem.
The configuration is relativly small so it doesn't consume a lot of
memory, while significantly reducing the query time whenever it is needed.

Also, the check for integration was only disabling calendar display if
necessarry, now it also enables it when integration is enabled.


Former-commit-id: 0f42679e74f3225269110aa91a8bd527eb18ecde